+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgraded allotropy python requirement to python 10
 - Updated ASM model class typing to use or union
 - Implement default value for sample role names in AppBio Quantstudio
+- Added kw_only=True for generated schema models
 
 ### Deprecated